+++ User-Service/src/main/java/org/caesar/userservice/Config/BearerTokenExtractor.java
@@ -1,6 +1,5 @@
 package org.caesar.userservice.Config;
 
-import com.nimbusds.jwt.JWTClaimsSet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.HttpStatus;
@@ -17,18 +16,24 @@ public class BearerTokenExtractor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+        //Presa del token dall'header della richiesta
         String authorizationHeader = request.getHeader("Authorization");
 
+        //Controllo che l'header non sia nulla e sia effettivamente un token bearer
         if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+
+            //Presa della stringa effettiva che rappresenta il token
             String bearerToken = authorizationHeader.substring(7);
             try {
-                // Estrarre il nome utente preferito dal token bearer
-                String preferredUsername = extractPreferredUsernameFromToken(bearerToken);
-                // Impostare il nome utente preferito nella richiesta
 
+                //Estrazione dell'username dal token
+                String preferredUsername = extractPreferredUsernameFromToken(bearerToken);
+
+                //Impostazione di un attributo della chiamata che rappresenta l'username dell'utente che ha effettuato la chiamata
                 request.setAttribute("preferred_username", preferredUsername);
             } catch (ParseException e) {
-                // Gestisci il token non valido
+
+                //Gestione di un token non valido
                 response.setStatus(HttpStatus.UNAUTHORIZED.value());
                 return false;
             }
@@ -37,8 +42,8 @@ public class BearerTokenExtractor implements HandlerInterceptor {
     }
 
     private String extractPreferredUsernameFromToken(String bearerToken) throws ParseException {
-        // Decodifica il token JWT e ottieni il campo preferred_username
-        JwtDecoder jwtDecoder= NimbusJwtDecoder.withJwkSetUri("http://25.24.244.170:8080/realms/CaesarRealm/protocol/openid-connect/certs").build(); // Sostituisci con il tuo issuer URI per JWK
+        //Decodifica del token e presa del claim preferred_username
+        JwtDecoder jwtDecoder= NimbusJwtDecoder.withJwkSetUri("http://25.24.244.170:8080/realms/CaesarRealm/protocol/openid-connect/certs").build();
         Jwt jwt = jwtDecoder.decode(bearerToken);
 
         return jwt.getClaim("preferred_username");
