+++ User-Service/src/main/java/org/caesar/userservice/Controller/SearchController.java
@@ -1,0 +1,13 @@
+package org.caesar.userservice.Controller;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/user-api")
+@RequiredArgsConstructor
+@Slf4j
+public class SearchController {
+}
