+++ User-Service/src/main/java/org/caesar/userservice/Data/Services/ProfilePicService.java
@@ -4,6 +4,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 public interface ProfilePicService {
 
-    boolean saveImage(MultipartFile file);
+    boolean saveImage(String username, MultipartFile file);
     byte[] getUserImage(String username);
 }
