+++ User-Service/src/main/java/org/caesar/userservice/Data/Services/UserService.java
@@ -1,16 +1,13 @@
 package org.caesar.userservice.Data.Services;
 
 import org.caesar.userservice.Dto.UserDTO;
-import org.caesar.userservice.Dto.UserIdDTO;
 import org.caesar.userservice.Dto.UserRegistrationDTO;
-import reactor.core.publisher.Mono;
 
 import java.util.List;
 
 
 public interface UserService {
 
-    UserIdDTO getUserId();
     UserDTO getUser(String username);
     List<UserDTO> getUsers(int start);
     List<String> getUsersByUsername(String username);
